images/image_sorting/pyproject.toml
[project]
name = "image-sorting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mutagen>=1.47.0",
    "piexif>=1.1.3",
    "pillow>=11.2.1",
    "pywin32>=310",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
]
